+++ zenodo-plugin.php
@@ -13,10 +13,10 @@
  * @package           Display_Your_Zenodo_Community
  *
  * @wordpress-plugin
- * Plugin Name:       Display your Zenodo Community
+ * Plugin Name:       Display your Zenodo Publications
  * Plugin URI:        https://github.com/dariah-eric/display-your-zenodo-community
  * Description:       This plugin allows user to view their community data from Zenodo on WordPress
- * Version:           1.0.3
+ * Version:           1.1.0
  * Author:            DARIAH-EU
  * Author URI:        https://www.dariah.eu
  * License:           Apache License - 2.0